+++ src/components/students/index.ts
@@ -1,0 +1,2 @@
+export * from './class-sessions';
+export * from './payments';
